+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Other
 
 - Improves error messages when passing components types as `(A, B)` instead of the required `Val.((A, B))` (#191)
+- Checks for duplicate components on query construction (#255)
 
 ## [[v0.1.0]](https://github.com/mlange-42/Ark.jl/tree/v0.1.0)